backend/requirements.txt
quart==0.19.8
quart-cors==0.7.0
hypercorn==0.17.3
openai==1.12.0
chromadb==0.4.22
python-dotenv==1.0.1
PyPDF2==3.0.1
langchain==0.1.9
numpy==1.26.4
scikit-learn==1.5.2  # Updated to latest stable version